+++ tidy3d/components/data/dataset.py
@@ -11,6 +11,7 @@ import pydantic as pd
 from .data_array import DataArray
 from .data_array import ScalarFieldDataArray, ScalarFieldTimeDataArray, ScalarModeFieldDataArray
 from .data_array import ModeIndexDataArray
+from .data_array import TriangleMeshDataArray
 
 from ..base import Tidy3dBaseModel
 from ..types import Axis
@@ -380,4 +381,15 @@ class PermittivityDataset(AbstractFieldDataset):
         ...,
         title="Epsilon zz",
         description="Spatial distribution of the zz-component of the relative permittivity.",
+    )
+
+
+class TriangleMeshDataset(Dataset):
+    """Dataset for storing triangular surface data."""
+
+    surface_mesh: TriangleMeshDataArray = pd.Field(
+        ...,
+        title="Surface mesh data",
+        description="Dataset containing the surface triangles and corresponding face indices "
+        "for a surface mesh.",
     )